moved draw game to display manager
