+++ static/modules/displaymanager.js
@@ -1,6 +1,11 @@
-import { Vector } from "./vector";
+import { Vector } from "./vector.js";
+import { getCameraEntity } from "./gamemanager.js";
+import { Entity } from "./entity.js";
+import { GuiElement } from "./guielement.js";
 
 const BLUR_SCALAR = 2;
+
+export const SPLATTER_SCALAR = 4;
 
 class DisplayManager {
   /** @type {Vector} */
@@ -55,6 +60,127 @@ class DisplayManager {
     };
 
     this.displayCanvas.addEventListener("fullscreenchange", exitHandler, false);
+  }
+
+  /**
+   * draw all the entities and particles
+   * @param {Entity[]} entities
+   * @param {Entity[]} particles
+   * @param {Map<string, GuiElement>} guiElements
+   */
+  drawGame(entities, particles, guiElements) {
+    // reposition camera if there is a followed entity
+    const { width: screenWidth, height: screenHeight } = getScreenDimensions();
+    if (getCameraEntity() !== undefined) {
+      setCameraOffset(
+        getCameraEntity()
+          .drawPos.mult(-1)
+          .add(new Vector(screenWidth / 2, screenHeight / 2))
+      );
+    }
+
+    // clear the display canvas with black rectangle
+    this.displayContext.fillRect(
+      0,
+      0,
+      this.displayCanvas.width,
+      this.displayCanvas.height
+    );
+
+    // clear the drawing canvas with alpha 0
+    this.context.clearRect(0, 0, this.canvas.width, this.canvas.height);
+
+    // clear the blur canvas with alpha 0
+    this.blurContext.clearRect(
+      0,
+      0,
+      this.blurCanvas.width,
+      this.blurCanvas.height
+    );
+
+    // copy the splatter canvas onto the drawing canvas
+    const targetCanvas = this.displayCanvas;
+    const targetContext = this.displayContext;
+    const splatterVec = getCameraOffset().mult(-1 / SPLATTER_SCALAR);
+    const displayRatio = this.canvas.width / targetCanvas.width;
+    targetContext.drawImage(
+      this.splatterCanvas,
+      splatterVec.x,
+      splatterVec.y,
+      (targetCanvas.width / SPLATTER_SCALAR) * displayRatio,
+      (targetCanvas.height / SPLATTER_SCALAR) * displayRatio,
+      0,
+      0,
+      targetCanvas.width,
+      targetCanvas.height
+    );
+
+    // save drawing context
+    this.context.save();
+    // run draw func specified by game programmer
+    this.drawFunc();
+
+    // draw all particles
+    for (let i = 0; i < particles.length; i++) {
+      if (particles[i].onScreen()) {
+        particles[i].draw();
+      }
+    }
+
+    // draw all entities
+    for (let i = 0; i < entities.length; i++) {
+      if (entities[i].onScreen()) {
+        entities[i].draw();
+      }
+    }
+    // restore drawing context
+    this.context.restore();
+
+    // copy the drawing canvas onto the blur canvas
+    this.blurContext.drawImage(
+      this.canvas,
+      0,
+      0,
+      this.canvas.width / BLUR_SCALAR,
+      this.canvas.height / BLUR_SCALAR
+    );
+
+    // TODO move to DisplayManager
+    // align camera, draw the gui, reset camera
+    // this is after the draw canvas is copied to the blur canvas
+    // move this to before if you want the gui blurred
+    const originalOffset = this.cameraOffset;
+    this.cameraOffset = new Vector(0, 0);
+    for (const guiKey of guiElements.keys()) {
+      if (guiElements.get(guiKey).active) {
+        guiElements.get(guiKey).draw();
+      }
+    }
+    this.cameraOffset = originalOffset;
+
+    // copy the blur canvas onto the display canvas
+    this.displayContext.drawImage(
+      this.blurCanvas,
+      0,
+      0,
+      this.displayCanvas.width,
+      this.displayCanvas.height
+    );
+
+    // save display context
+    this.displayContext.save();
+    this.displayContext.globalCompositeOperation = "lighter";
+    // copy the drawing canvas onto the display canvas
+    this.displayContext.drawImage(
+      this.canvas,
+      0,
+      0,
+      this.displayCanvas.width,
+      this.displayCanvas.height
+    );
+
+    // restore display context
+    this.displayContext.restore();
   }
 }
 
@@ -140,7 +266,6 @@ export function setCameraOffset(cameraOffset) {
   displayManager.cameraOffset = cameraOffset;
 }
 
-// TODO move this to DisplayManager
 /**
  * @returns {Promise<void>}
  */
@@ -154,4 +279,14 @@ export function toggleFullscreen() {
     // exit fullscreen
     return document.exitFullscreen();
   }
+}
+
+/**
+ * draw all the entities and particles
+ * @param {Entity[]} entities
+ * @param {Entity[]} particles
+ * @param {Map<string, GuiElement>} guiElements
+ */
+export function drawGame(entities, particles, guiElements) {
+  displayManager.drawGame(entities, particles, guiElements);
 }
