+++ static/modules/gamemanager.js
@@ -25,6 +25,7 @@ import {
 // TODO move this
 const BLUR_SCALAR = 2;
 
+// TODO move this to displaymanager
 export const SPLATTER_SCALAR = 4;
 
 class GameManager {
@@ -286,122 +287,6 @@ class GameManager {
     );
   }
 
-  // TODO move to DisplayManager
-  drawGame() {
-    // reposition camera if there is a followed entity
-    const { width: screenWidth, height: screenHeight } = getScreenDimensions();
-    if (this.cameraEntity !== undefined) {
-      setCameraOffset(
-        this.cameraEntity.drawPos
-          .mult(-1)
-          .add(new Vector(screenWidth / 2, screenHeight / 2))
-      );
-    }
-
-    // clear the display canvas with black rectangle
-    this.displayContext.fillRect(
-      0,
-      0,
-      this.displayCanvas.width,
-      this.displayCanvas.height
-    );
-
-    // clear the drawing canvas with alpha 0
-    this.context.clearRect(0, 0, this.canvas.width, this.canvas.height);
-
-    // clear the blur canvas with alpha 0
-    this.blurContext.clearRect(
-      0,
-      0,
-      this.blurCanvas.width,
-      this.blurCanvas.height
-    );
-
-    // copy the splatter canvas onto the drawing canvas
-    const targetCanvas = this.displayCanvas;
-    const targetContext = this.displayContext;
-    const splatterVec = getCameraOffset().mult(-1 / SPLATTER_SCALAR);
-    const displayRatio = this.canvas.width / targetCanvas.width;
-    targetContext.drawImage(
-      this.splatterCanvas,
-      splatterVec.x,
-      splatterVec.y,
-      (targetCanvas.width / SPLATTER_SCALAR) * displayRatio,
-      (targetCanvas.height / SPLATTER_SCALAR) * displayRatio,
-      0,
-      0,
-      targetCanvas.width,
-      targetCanvas.height
-    );
-
-    // save drawing context
-    this.context.save();
-    // run draw func specified by game programmer
-    this.drawFunc();
-
-    // draw all particles
-    for (let i = 0; i < this.particles.length; i++) {
-      if (this.particles[i].onScreen()) {
-        this.particles[i].draw();
-      }
-    }
-
-    // draw all entities
-    for (let i = 0; i < this.entities.length; i++) {
-      if (this.entities[i].onScreen()) {
-        this.entities[i].draw();
-      }
-    }
-    // restore drawing context
-    this.context.restore();
-
-    // copy the drawing canvas onto the blur canvas
-    this.blurContext.drawImage(
-      this.canvas,
-      0,
-      0,
-      this.canvas.width / BLUR_SCALAR,
-      this.canvas.height / BLUR_SCALAR
-    );
-
-    // TODO move to DisplayManager
-    // align camera, draw the gui, reset camera
-    // this is after the draw canvas is copied to the blur canvas
-    // move this to before if you want the gui blurred
-    const originalOffset = this.cameraOffset;
-    this.cameraOffset = new Vector(0, 0);
-    for (const guiKey of this.guiElements.keys()) {
-      if (this.guiElements.get(guiKey).active) {
-        this.guiElements.get(guiKey).draw();
-      }
-    }
-    this.cameraOffset = originalOffset;
-
-    // copy the blur canvas onto the display canvas
-    this.displayContext.drawImage(
-      this.blurCanvas,
-      0,
-      0,
-      this.displayCanvas.width,
-      this.displayCanvas.height
-    );
-
-    // save display context
-    this.displayContext.save();
-    this.displayContext.globalCompositeOperation = "lighter";
-    // copy the drawing canvas onto the display canvas
-    this.displayContext.drawImage(
-      this.canvas,
-      0,
-      0,
-      this.displayCanvas.width,
-      this.displayCanvas.height
-    );
-
-    // restore display context
-    this.displayContext.restore();
-  }
-
   collideWithEntities() {
     // TODO why is this logic not part of the collision.js file?
     // generate the type map for faster access
@@ -523,7 +408,7 @@ class GameManager {
     this.performTween(this.particles, timeLeft);
     this.overTime = -timeLeft;
 
-    this.drawGame();
+    drawGame(this.entities, this.particles, this.guiElements);
     //this.destroyEntities();
 
     // increase the time
